add Procfile
Procfile is a file with configuration how to start the app. The app is
already built at this stage, so now we need to tell heroku how to start
it. This setup is very simple, a web type of heroku server (other types
are, for example, background jobs workers) should start a web server.

https://devcenter.heroku.com/articles/procfile